requirements.txt
asgiref==3.8.1
Django==5.0.14
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
